requirements.txt
# FastAPI & Server
fastapi
uvicorn

# PDF Processing
pypdf

# OpenAI & LLM
openai

# Environment Variables
python-dotenv

# Utilities
python-multipart

